backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
pydantic==2.11.2
pydantic-settings==2.9.1
python-jose[cryptography]==3.4.0
passlib==1.7.4
argon2-cffi==23.1.0
python-dotenv==1.1.0
pytest==8.3.5
httpx==0.28.1
